requirements.txt
fastapi
uvicorn[standard]
python-dotenv
stripe
requests
python-nmap
python-owasp-zap-v2.4

